Feature/hu2 solicitud prestamo

+++ infrastructure/driven-adapters/r2dbc-postgresql/src/main/java/co/pragma/r2dbc/repository/EstadoSolicitudReactiveRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.query.ReactiveQueryByExampleExecutor;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import reactor.core.publisher.Mono;
 
-public interface EstadoSolicitudReactiveRepository extends ReactiveCrudRepository<EstadoSolicitudEntity, Integer>, ReactiveQueryByExampleExecutor<EstadoSolicitudEntity> {
+public interface EstadoSolicitudReactiveRepository extends ReactiveCrudRepository<EstadoSolicitudEntity, Short>, ReactiveQueryByExampleExecutor<EstadoSolicitudEntity> {
 
     Mono<EstadoSolicitudEntity> findByNombre(String nombre);
 }
